feat(oauth2): add some checks on auth parameters
