+++ sample.yml
@@ -13,6 +13,7 @@ swagger:
 
     clientId: "demoapp"
     appName: "dropwizard-sample"
+    realm: "smoketurner"
     additionalQueryStringParams:
       foo: "bar"
